+++ clang/lib/AST/TextNodeDumper.cpp
@@ -1580,6 +1580,13 @@ void clang::TextNodeDumper::VisitDependentScopeDeclRefExpr(
   dumpNestedNameSpecifier(Node->getQualifier());
 }
 
+void clang::TextNodeDumper::VisitDependentTemplateIdExpr(
+    const DependentTemplateIdExpr *Node) {
+  OS << (Node->isConceptReference() ? " concept" : " variable template");
+  OS << ' ';
+  dumpBareTemplateName(Node->getTemplateName());
+}
+
 void TextNodeDumper::VisitUnresolvedLookupExpr(
     const UnresolvedLookupExpr *Node) {
   OS << " (";